Consolidate godexviewer and rustdexviewer
