+++ dexviewer/godexviewer/main.go
@@ -1,10 +1,8 @@
 package main
 
 import (
-	"bufio"
 	"bytes"
 	"fmt"
-	"strings"
 	"syscall/js"
 
 	"github.com/csnewman/dextk"
@@ -15,38 +13,11 @@ type DexFile struct {
 }
 
 var dexFiles = make(map[int]DexFile)
-var proguardMapping = make(map[string]string)
-
-func parseProguardMapping(content string) map[string]string {
-	mapping := make(map[string]string)
-	scanner := bufio.NewScanner(strings.NewReader(content))
-	for scanner.Scan() {
-		line := scanner.Text()
-		if strings.HasSuffix(line, ":") && strings.Contains(line, " -> ") {
-			parts := strings.Split(line, " -> ")
-			originalName := parts[0]
-			obfuscatedName := strings.TrimSuffix(parts[1], ":")
-			mapping[obfuscatedName] = originalName
-		}
-	}
-	return mapping
-}
-
-func loadProguardMapping(this js.Value, args []js.Value) any {
-	mappingContent := args[0].String()
-	proguardMapping = parseProguardMapping(mappingContent)
-	fmt.Println("Loaded ProGuard mapping for", len(proguardMapping), "classes")
-	return nil
-}
 
 func main() {
 	// Create dextk object to store functions
 	dextkObj := map[string]interface{}{
-		"createDex":             js.FuncOf(createDex),
-		"getClasses":            js.FuncOf(getClasses),
-		"getMethodNames":        js.FuncOf(getMethodNames),
 		"getMethodInstructions": js.FuncOf(getMethodInstructions),
-		"loadProguardMapping":   js.FuncOf(loadProguardMapping),
 	}
 
 	// Export dextk object to JavaScript
@@ -56,228 +27,48 @@ func main() {
 	<-make(chan bool)
 }
 
-func createDex(this js.Value, args []js.Value) any {
+func getMethodInstructions(this js.Value, args []js.Value) any {
 	// Get the Uint8Array from JavaScript
 	array := args[0]
+	// Get the class ID from JavaScript
+	classId := args[1].Int()
+	// Get the method name from JavaScript
+	methodName := args[2].String()
 
 	// Create Go byte slice and copy data
 	dexbytes := make([]byte, array.Length())
 	js.CopyBytesToGo(dexbytes, array)
-
-	// Generate next available key
-	key := len(dexFiles)
-
-	// Store in map
-	dexFiles[key] = DexFile{
-		bytes: dexbytes,
-	}
-
-	return key
-}
-
-func getClasses(this js.Value, args []js.Value) any {
-	// Get the dex file ID from JavaScript
-	dexId := args[0].Int()
-
-	// Get the dex file from map
-	dexFile, ok := dexFiles[dexId]
-	if !ok {
-		fmt.Println("Invalid dex file ID")
-		return nil
-	}
-
-	r, err := dextk.Read(bytes.NewReader(dexFile.bytes))
-	if err != nil {
-		fmt.Println(err)
-	}
-
-	var classInfos []any
-
-	ci := r.ClassIter()
-	for ci.HasNext() {
-		node, err := ci.Next()
-		if err != nil {
-			fmt.Println(err)
-			continue
-		}
-
-		obfuscatedName := node.Name.String()
-		originalName, ok := proguardMapping[obfuscatedName]
-		if !ok {
-			originalName = obfuscatedName
-		}
-
-		classInfos = append(classInfos, map[string]any{
-			"obfuscated": obfuscatedName,
-			"original":   originalName,
-		})
-	}
-
-	fmt.Println("Done")
-	return js.ValueOf(classInfos)
-}
-
-func getMethodNames(this js.Value, args []js.Value) any {
-	// Get the dex file ID from JavaScript
-	dexId := args[0].Int()
-
-	// Get the class name from JavaScript
-	className := args[1].String()
-
-	// Get the dex file from map
-	dexFile, ok := dexFiles[dexId]
-	if !ok {
-		fmt.Println("Invalid dex file ID")
-		return nil
-	}
-
-	r, err := dextk.Read(bytes.NewReader(dexFile.bytes))
-	if err != nil {
-		fmt.Println(err)
-		return nil
-	}
-
-	var methodNames []string
-
-	// Iterate through classes to find matching class
-	ci := r.ClassIter()
-	for ci.HasNext() {
-		node, err := ci.Next()
-		if err != nil {
-			fmt.Println(err)
-			continue
-		}
-
-		if node.Name.String() == className {
-			// Add direct methods
-			for _, method := range node.DirectMethods {
-				signature := formatMethod(method)
-				methodNames = append(methodNames, signature)
-			}
-
-			// Add virtual methods
-			for _, method := range node.VirtualMethods {
-				signature := formatMethod(method)
-				methodNames = append(methodNames, signature)
-			}
-			break
-		}
-	}
-
-	return js.ValueOf(stringSliceToAnySlice(methodNames))
-}
-
-func formatMethod(method dextk.MethodNode) string {
-	shorty := method.Shorty.String()
-	var returnType string
-	switch shorty[0] {
-	case 'V':
-		returnType = "void"
-	case 'Z':
-		returnType = "boolean"
-	case 'B':
-		returnType = "byte"
-	case 'S':
-		returnType = "short"
-	case 'C':
-		returnType = "char"
-	case 'I':
-		returnType = "int"
-	case 'J':
-		returnType = "long"
-	case 'F':
-		returnType = "float"
-	case 'D':
-		returnType = "double"
-	case 'L':
-		returnType = method.ReturnType.String()
-	case '[':
-		returnType = method.ReturnType.String() + "[]"
-	default:
-		returnType = "unknown"
-	}
-
-	// Format parameter types
-	var params []string
-	for _, c := range shorty[1:] {
-		var paramType string
-		switch c {
-		case 'Z':
-			paramType = "boolean"
-		case 'B':
-			paramType = "byte"
-		case 'S':
-			paramType = "short"
-		case 'C':
-			paramType = "char"
-		case 'I':
-			paramType = "int"
-		case 'J':
-			paramType = "long"
-		case 'F':
-			paramType = "float"
-		case 'D':
-			paramType = "double"
-		case 'L':
-			paramType = "Object"
-		case '[':
-			paramType = "Array"
-		default:
-			paramType = "unknown"
-		}
-		params = append(params, paramType)
-	}
-
-	return fmt.Sprintf("%s %s(%s)", returnType, method.Name.String(), strings.Join(params, ", "))
-}
-
-func getMethodInstructions(this js.Value, args []js.Value) any {
-	// Get the dex file ID from JavaScript
-	dexId := args[0].Int()
-	// Get the class name from JavaScript
-	className := args[1].String()
-	// Get the method name from JavaScript
-	methodName := args[2].String()
-
-	// Get the dex file from map
-	dexFile, ok := dexFiles[dexId]
-	if !ok {
-		fmt.Println("Invalid dex file ID")
-		return nil
-	}
-
-	r, err := dextk.Read(bytes.NewReader(dexFile.bytes))
+	r, err := dextk.Read(bytes.NewReader(dexbytes))
 	if err != nil {
 		fmt.Println(err)
 		return nil
 	}
 
 	// Find the method in the class
-	ci := r.ClassIter()
-	for ci.HasNext() {
-		node, err := ci.Next()
-		if err != nil {
-			fmt.Println(err)
-			continue
-		}
+	class, err := r.ReadClassAndParse(uint32(classId))
+	if err != nil {
+		fmt.Println(err)
+		return nil
+	}
 
-		if node.Name.String() == className {
-			// Check direct methods
-			for _, method := range node.DirectMethods {
-				if method.Name.String() == methodName {
-					return js.ValueOf(stringSliceToAnySlice(getInstructions(r, method)))
-				}
-			}
-			// Check virtual methods
-			for _, method := range node.VirtualMethods {
-				if method.Name.String() == methodName {
-					return js.ValueOf(stringSliceToAnySlice(getInstructions(r, method)))
-				}
-			}
-			break
+	fmt.Println("Class id:", classId, "Class name:", class.Name.String())
+
+	// Check direct methods
+	for _, method := range class.DirectMethods {
+		fmt.Println("Direct method:", method.Name.String(), "Method name:", methodName)
+		if method.Name.String() == methodName {
+			return js.ValueOf(stringSliceToAnySlice(getInstructions(r, method)))
+		}
+	}
+	// Check virtual methods
+	for _, method := range class.VirtualMethods {
+		fmt.Println("Virtual method:", method.Name.String(), "Method name:", methodName)
+		if method.Name.String() == methodName {
+			return js.ValueOf(stringSliceToAnySlice(getInstructions(r, method)))
 		}
 	}
 
+	fmt.Println("Method not found")
 	return nil
 }
