+++ package.json
@@ -14,9 +14,8 @@
         "mhtml",
         "classfile",
         "browser",
-        "dexviewer",
         "textviewer",
-        "rustdexviewer",
+        "dexviewer",
         "main",
         "webgl_previewer",
         "jqviewer",
