--- rustdexviewer/godexviewer/go.mod
@@ -1,5 +1,0 @@
-module dexviewer
-
-go 1.23.4
-
-require github.com/csnewman/dextk v0.3.0 // indirect
