--- rustdexviewer/package.json
@@ -1,18 +1,0 @@
-{
-  "name": "wasmtools-rustdexviewer",
-  "version": "1.0.0",
-  "author": "Maurice Lam",
-  "scripts": {
-    "build": "node build.mjs",
-    "watch": "BUILD_MODE=dev node build.mjs"
-  },
-  "license": "MIT",
-  "description": "",
-  "devDependencies": {
-    "esbuild-plugin-wasm": "^1.1.0"
-  },
-  "dependencies": {
-    "react": "^18.3.1",
-    "react-dom": "^18.3.1"
-  }
-}
